pyproject.toml
[project]
name = "py-web-automation"
version = "3.2.0"
description = "Comprehensive Python framework for automated testing of web applications"
authors = [{ name = "Ravil Shakerov", email = "xellaopromaster@yandex.ru" }]
requires-python = ">=3.12"
license = { text = "MIT" }
readme = "README.md"
dependencies = [
    # Core runtime dependencies
    "httpx>=0.27.0", # HTTP client for API testing (REST, GraphQL, SOAP)
    "playwright>=1.48.0", # Browser automation for UI testing
    "msgspec>=0.19.0", # High-performance data validation and serialization
    "loguru>=0.7.2", # Logging framework
    "pyyaml>=6.0.1", # YAML configuration support
    "websockets>=13.1", # WebSocket client support
    "aiosqlite>=0.20.0", # SQLite database adapter (async)
    "pillow>=12.0.0",
]

keywords = [
    "web automation",
    "api testing",
    "ui testing",
    "browser automation",
    "testing framework",
    "rest api",
    "graphql",
    "grpc",
    "soap",
    "websocket",
    "database testing",
    "playwright",
    "httpx",
    "async",
]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]


[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["py_web_automation"]

[tool.pytest.ini_options]
# Test discovery
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# Add project root to Python path for imports
pythonpath = ["."]

# Output and reporting
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--tb=short",
    "--cov=py_web_automation",
    "--cov-report=html:report/coverage",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
    "--durations=10",
    "--maxfail=5",
    "-n", "8",
]

# Markers
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "performance: Performance tests",
    "api: API testing tests",
    "ui: UI testing tests",
    "db: Database testing tests",
    "websocket: WebSocket testing tests",
    "grpc: gRPC testing tests",
    "soap: SOAP testing tests",
    "graphql: GraphQL testing tests",
    "smoke: Smoke tests",
    "regression: Regression tests",
]

# Test execution
minversion = "8.0"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["py_web_automation"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.coverage.html]
directory = "report/coverage"

[dependency-groups]
# Testing dependencies
test = [
    "pytest>=8.5.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
    "pytest-playwright>=0.4.4",
    "allure-pytest>=2.13.2",
]

# Development dependencies (linting, type checking, code quality)
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.2.0",
    "prospector>=1.17.3",
    "pydeps>=3.0.1",
    "pyflakes>=3.4.0",
    "radon>=6.0.1",
    "vulture>=2.14",
    "types-pyyaml>=6.0.12.20250516",
    "solid-checker>=1.0.4",
    "mypy>=1.19.0",
]

# All development tools (test + dev)
# Note: Use 'uv sync --all-groups' to install all dependency groups
# Or use: 'uv sync --group test --group dev'
all = [
    # Testing dependencies
    "pytest>=8.5.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.8.0",
    "pytest-random-order>=1.2.0",
    "pytest-playwright>=0.4.4",
    "allure-pytest>=2.13.2",
    # Development dependencies
    "ruff>=0.8.0",
    "mypy>=1.19.0",
    "pre-commit>=4.2.0",
    "prospector>=1.17.3",
    "pydeps>=3.0.1",
    "pyflakes>=3.4.0",
    "radon>=6.0.1",
    "vulture>=2.14",
    "types-pyyaml>=6.0.12.20250516",
    "solid-checker>=1.0.4",
    "asyncpg>=0.29.0",
    "aiomysql>=0.2.0",
    "grpclib>=0.4.7",
    "zeep>=4.2.1",
    "graphql-core>=3.2.3",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "B027",  # Empty method in abstract base class (optional hooks in Plugin)
    "B024",  # Abstract base class without abstract methods (Component is a marker class)
    "B017",  # Do not assert blind exception (database-specific errors vary by backend)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
warn_return_any = false  # Disable for dynamic JSON parsing
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
explicit_package_bases = true
ignore_missing_imports = false
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "playwright.*",
    "websockets.*",
    "zeep.*",
    "grpclib.*",
    "graphql.*",
    "asyncpg.*",
    "aiomysql.*",
    "aiosqlite.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "examples.*",
]
# Relax type checking for tests and examples
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = false
warn_return_any = false
warn_unused_ignores = false
ignore_errors = true
